backend/ubi_project/requirements.txt
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
Django==3.1.3
django-countries==6.1.3
django-environ==0.4.5
django-pgviews==0.5.7
djangorestframework==3.12.2
flake8==3.8.4
httpie==2.3.0
idna==2.10
mccabe==0.6.1
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.7.2
PySocks==1.7.1
pytz==2020.4
requests==2.25.0
requests-toolbelt==0.9.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
